+++ tests/test_generate_molecules.py
@@ -8,10 +8,12 @@ from rdkit import Chem
 from visual_graph_datasets.visualization.base import create_frameless_figure
 from visual_graph_datasets.processing.molecules import MoleculeProcessing
 
+from vgd_counterfactuals.base import CounterfactualGenerator
 from vgd_counterfactuals.generate.molecules import get_neighborhood
 from vgd_counterfactuals.generate.molecules import get_free_valence_map
 from vgd_counterfactuals.generate.molecules import get_valid_bond_additions
 from vgd_counterfactuals.generate.molecules import get_valid_bond_removals
+from vgd_counterfactuals.generate.molecules import is_bridge_head_carbon
 from .util import ARTIFACTS_PATH
 
 
@@ -19,10 +21,7 @@ def test_get_neighborhood():
 
     smiles = 'CCCCCC'
 
-    start_time = time.time()
     neighbors = get_neighborhood(smiles)
-    end_time = time.time()
-    print(f'generated {len(neighbors)} molecular mutations in {end_time - start_time:.3f} seconds')
     assert len(neighbors) != 0
 
     processing = MoleculeProcessing()
@@ -34,6 +33,11 @@ def test_get_neighborhood():
             ax.imshow(image)
             pdf.savefig(fig)
             plt.close(fig)
+
+    smiles = 'O=O'
+    neighbors = get_neighborhood(smiles)
+    assert len(neighbors) != 0
+    print(neighbors)
 
 
 def test_get_valid_bond_additions():
@@ -64,5 +68,34 @@ def test_get_valid_bond_removals():
     result = get_valid_bond_removals(mol)
     assert len(result) != 0
     assert len(result) == 3
+
+
+def test_bridgehead_carbons_exclusion():
+    """
+    15.05.23 - Tests a problem which Hunter has noticed about the generation of molecular counterfactuals.
+    He pointed out a specific molecular configuration which RDkit technically says is valid solely based
+    on the valence rules etc., but which never practically happens in chemistry. Tests a filter which
+    detects this kind of pattern and excludes it from the neighborhood of an atom.
+    """
+    # This molecule contains the structure which is not supposed to happen and should therefore be
+    # detected by the filter.
+    smiles = 'NCCCC1=CC2=CC=C1C2'
+    mol = Chem.MolFromSmiles(smiles)
+
+    processing = MoleculeProcessing()
+    fig, node_positions = processing.visualize_as_figure(smiles, width=1000, height=1000)
+    fig.savefig(os.path.join(ARTIFACTS_PATH, 'bridgehead_carbons.pdf'))
+
+    match = is_bridge_head_carbon(mol)
+    assert match is True
+
+    # Now this is molecule contains only a normal carbon ring and should therefore not be detected by
+    # the filter!
+    smiles = 'NCCCCC1=CC=CC=C1'
+    mol = Chem.MolFromSmiles(smiles)
+
+    match = is_bridge_head_carbon(mol)
+    assert match is False
+