+++ HISTORY.rst
@@ -6,3 +6,16 @@ Changelog
 
 Initial version
 
+
+0.1.1 - 10.05.2023
+------------------
+
+- Fixed a crucial bug in the generation of chemical counterfactuals where single atom "molecules" were
+  allowed and
+
+0.1.2 - 15.05.2023
+------------------
+
+- Added the possibility to add additional filters to the generation of a molecular neighborhood which
+  exclude certain patterns for the resulting neighbors.
+- Added a default molecular filter which excludes bridge head carbon structures from appearing